bug: cart no longer showing after submit checkout
